requirements.txt
Flask
colorama
cryptography
pyserial
pytz
